refactor: move template engine integration tests to engine crate
