+++ dan_layer/engine/Cargo.toml
@@ -31,4 +31,5 @@ tempfile = "3.3.0"
 wasmer = "2.3.0"
 wasmer-middlewares = "2.3.0"
 
-
+[dev-dependencies]
+tari_template_test_tooling = { path = "../template_test_tooling" }
